Refactored TotalFlashcardBrowser Interface
